Fix data type for rel_l2_norm_sum in evaluate_channel_metrics
Changed the data type of rel_l2_norm_sum from float64 to float32 for improved performance and memory efficiency. Other related tensors remain as float64 to maintain precision where necessary.

+++ src/sweep/metrics.py
@@ -58,7 +58,7 @@ def evaluate_channel_metrics(
 
             if rel_l2_norm_sum is None:
                 channels = rel_l2_norm.shape[1]
-                rel_l2_norm_sum = torch.zeros(channels, dtype=torch.float64, device=device)
+                rel_l2_norm_sum = torch.zeros(channels, dtype=torch.float32, device=device)
                 rel_l2_denorm_sum = torch.zeros(channels, dtype=torch.float64, device=device)
                 mse_norm_sum = torch.zeros(channels, dtype=torch.float64, device=device)
                 mse_denorm_sum = torch.zeros(channels, dtype=torch.float64, device=device)
